Refactor: Extract tests from God Objects and fix clippy warnings
- Extracted tests from `src/db/mod.rs` to `src/db/tests.rs` to reduce file size.
- Extracted tests from `src/query/parser/mod.rs` to `src/query/parser/tests.rs` to improve readability.
- Extracted `resolve_hnsw_config` helper method in `src/db/mod.rs` to flatten logic.
- Fixed `clippy::collapsible_if` warnings in `src/experimental/semantic_pathfinding.rs`.
- Fixed duplicated `#[cfg(test)]` attributes.

Verified with `cargo check`, `cargo fmt`, `cargo clippy`, and `cargo test`.
